+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [0.10.0] - 2019-10-05
+
+### Changed
+- fixed: critical bug: componentDidUpdate keeps re-rendering table (issue [#82](https://github.com/ngduc/react-tabulator/issues/82))
+
 ## [0.9.6] - 2019-09-06
 
 ### Added